The_Final_version_of_the_Project

+++ An_online_Cafe/order.py
@@ -2,16 +2,12 @@
 
 
 class Order():
-    def __init__(self,order_name:str,price:int,quantity:int) -> None:
+    def __init__(self,order_name:str,price:float,quantity:int) -> None:
         self.order_name=order_name
         self.price=price
         self.quantity=quantity
 
 
-
-
-
-# order1=Order('latte',coffe_menu['latte'],2)
